+++ tests/test_adapter.py
@@ -28,7 +28,6 @@ from .test_adapter_composition import ParallelAdapterInferenceTestMixin, Paralle
 from .test_adapter_conversion import ModelClassConversionTestMixin
 from .test_adapter_fusion_common import AdapterFusionModelTestMixin
 from .test_adapter_heads import PredictionHeadModelTestMixin
-from .test_adapter_setup_context import AdapterSetupContextTestMixin
 from .test_adapter_training import AdapterTrainingTestMixin
 
 
@@ -95,7 +94,6 @@ class BertAdapterTest(
     AdapterModelTestMixin,
     AdapterFusionModelTestMixin,
     PredictionHeadModelTestMixin,
-    AdapterSetupContextTestMixin,
     AdapterTrainingTestMixin,
     ParallelAdapterInferenceTestMixin,
     ParallelTrainingMixin,
@@ -130,7 +128,6 @@ class RobertaAdapterTest(
     AdapterModelTestMixin,
     AdapterFusionModelTestMixin,
     PredictionHeadModelTestMixin,
-    AdapterSetupContextTestMixin,
     ParallelAdapterInferenceTestMixin,
     RobertaAdapterTestBase,
     unittest.TestCase,
@@ -181,7 +178,6 @@ class DistilBertAdapterTest(
     EmbeddingTestMixin,
     AdapterFusionModelTestMixin,
     PredictionHeadModelTestMixin,
-    AdapterSetupContextTestMixin,
     AdapterTrainingTestMixin,
     ParallelAdapterInferenceTestMixin,
     ParallelTrainingMixin,
@@ -221,7 +217,6 @@ class BartAdapterTest(
     AdapterFusionModelTestMixin,
     EmbeddingTestMixin,
     PredictionHeadModelTestMixin,
-    AdapterSetupContextTestMixin,
     AdapterTrainingTestMixin,
     ParallelAdapterInferenceTestMixin,
     ParallelTrainingMixin,
@@ -259,7 +254,6 @@ class MBartAdapterTest(
     AdapterModelTestMixin,
     AdapterFusionModelTestMixin,
     PredictionHeadModelTestMixin,
-    AdapterSetupContextTestMixin,
     ParallelAdapterInferenceTestMixin,
     MBartAdapterTestBase,
     unittest.TestCase,
@@ -295,7 +289,6 @@ class GPT2AdapterTest(
     EmbeddingTestMixin,
     AdapterFusionModelTestMixin,
     PredictionHeadModelTestMixin,
-    AdapterSetupContextTestMixin,
     AdapterTrainingTestMixin,
     ParallelAdapterInferenceTestMixin,
     ParallelTrainingMixin,
